Fixed pearling snitches not firing when pearled into
PlayerTeleportEvent is fired before the player is actually teleported. As a result of this the location of the player as far as the event is concerned is where they are before they teleport. So when the snitch entry is handled it will take that location & not where they move to. By using the getTo() method & a player location override we can trigger snitches for when the player teleports in or out of a snitch field.